+++ src/main/java/com/example/demo/service/SourceService.java
@@ -9,7 +9,9 @@ import com.example.demo.repository.SheetRepository;
 import com.example.demo.repository.SourceRepository;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.Cacheable;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -20,16 +22,21 @@ import java.util.Optional;
 public class SourceService {
     private final SourceRepository sourceRepository;
     private final SheetRepository sheetRepository;
-    public SourceService(SourceRepository sourceRepository, SheetRepository sheetRepository){
+    private final JwtService jwtService;
+    public SourceService(SourceRepository sourceRepository, SheetRepository sheetRepository, JwtService jwtService){
         this.sourceRepository = sourceRepository;
         this.sheetRepository = sheetRepository;
+        this.jwtService = jwtService;
     }
 
     @CacheEvict(value = {"sourcesBySheet", "allSources"}, allEntries = true)
-    public SourceDTO create(SourceRequest request) {
-
+    public SourceDTO create(SourceRequest request, String auth) {
         Sheet sheet = sheetRepository.findById(request.getSheetId())
                 .orElseThrow(() -> new RuntimeException("Sheet not found"));
+        Integer userId = jwtService.extractUserId(auth);
+        if(!sheet.getUser().getId().equals(userId)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
 
         Source source = Source.builder()
                 .type(Type.valueOf(request.getType()))
@@ -45,14 +52,16 @@ public class SourceService {
     }
 
     @CacheEvict(value = {"sourcesBySheet", "allSources"}, allEntries = true)
-    public Optional<SourceDTO> update(Integer id, SourceRequest updated) {
-
+    public Optional<SourceDTO> update(Integer id, SourceRequest updated, String auth) {
+        Integer userId = jwtService.extractUserId(auth);
         Optional<Source> optional = sourceRepository.findById(id);
 
         if (optional.isEmpty()) return Optional.empty();
 
         Source source = optional.get();
-
+        if(!source.getSheet().getUser().getId().equals(userId)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         source.setType(Type.valueOf(updated.getType()));
         source.setAmount(updated.getAmount());
         source.setDescription(updated.getDescription());
@@ -63,16 +72,25 @@ public class SourceService {
     }
 
     @CacheEvict(value = {"sourcesBySheet", "allSources"}, allEntries = true)
-    public void delete(Integer id) {
+    public void delete(Integer id, String auth) {
         Source source = sourceRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Source not found"));
 
-        Integer sheetId = source.getSheet().getId();
+        Integer userId = jwtService.extractUserId(auth);
+        if(!source.getSheet().getUser().getId().equals(userId)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         sourceRepository.deleteById(id);
     }
 
     @Cacheable(value = "sourcesBySheet", key = "#sheetId")
-    public List<SourceDTO> getSourcesFromSheetId(Integer sheetId){
+    public List<SourceDTO> getSourcesFromSheetId(Integer sheetId, String auth){
+        Integer userId = jwtService.extractUserId(auth);
+        sheetRepository.findById(sheetId).ifPresentOrElse((s) -> {
+            if(!s.getUser().getId().equals(userId)){
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+            }
+        }, () -> {throw new ResponseStatusException(HttpStatus.NOT_FOUND);});
         return sourceRepository
                 .findAll()
                 .stream()
@@ -92,7 +110,11 @@ public class SourceService {
     }
 
     @Cacheable(value = "allSources")
-    public List<SourceDTO> getAllSources() {
+    public List<SourceDTO> getAllSources(String auth) {
+        String role = jwtService.extractUserRole(auth);
+        if(!role.equalsIgnoreCase("admin")){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         return sourceRepository.findAll().stream().map(this::mapToDTO).toList();
     }
 }
